pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "juntagrico-contribution"
description = "Create Contribution Rounds"
authors = [
    {name = "juntagrico", email = "python@juntagrico.org"},
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
dependencies = [
    "django-admin-sortable2~=2.1.10",
    "juntagrico>=1.7.0",
]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Environment :: Web Environment',
    'Framework :: Django',
    'Framework :: Django :: 4.2',
    'Intended Audience :: Developers',
    'Intended Audience :: Other Audience',
    'License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)',
    'Operating System :: OS Independent',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Topic :: Internet :: WWW/HTTP :: Site Management',
]
dynamic = ["version"]

[project.urls]
Homepage = "https://www.juntagrico.org"
Issues = "https://github.com/juntagrico/juntagrico-contribution/issues"

[tool.setuptools.packages.find]
include = ["juntagrico_contribution*"]

[tool.setuptools.dynamic]
version = {attr = "juntagrico_contribution.__version__"}

[tool.ruff.lint]
select = ["E", "F", "B"]
ignore = ["E501"]
